// src/generated-api/schemas/hackathonResponseHackathonState.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type HackathonResponseHackathonState = typeof HackathonResponseHackathonState[keyof typeof HackathonResponseHackathonState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const HackathonResponseHackathonState = {
  Pending: 'Pending',
  Ongoing: 'Ongoing',
  Ended: 'Ended',
} as const;