// src/generated-api/schemas/hackathonResponseType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type HackathonResponseType = typeof HackathonResponseType[keyof typeof HackathonResponseType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const HackathonResponseType = {
  WRITING: 'WRITING',
  AUDIO: 'AUDIO',
  ART: 'ART',
  BUSINESS: 'BUSINESS',
} as const;